+++ tests/ut/cpp/abstract/utils_test.cc
@@ -29,23 +29,23 @@ class TestUtils : public UT::Common {
 
 TEST_F(TestUtils, test_join) {
   // AbstractScalar
-  AbstractBasePtr abs_s1 = FromValue(1, false);
-  AbstractBasePtr abs_s2 = FromValue(2, false);
-  AbstractBasePtr abs_s_anything = FromValue(2, true);
+  AbstractBasePtr abs_s1 = FromValue(static_cast<int64_t>(1), false);
+  AbstractBasePtr abs_s2 = FromValue(static_cast<int64_t>(2), false);
+  AbstractBasePtr abs_s_anything = FromValue(static_cast<int64_t>(2), true);
 
   AbstractBasePtr res_s1 = abs_s1->Join(abs_s2);
   ASSERT_EQ(*res_s1, *abs_s_anything);
 
   // AbstractTuple join;
-  std::vector<int> list1 = {1, 2, 3, 4, 5};
-  std::vector<int> list2 = {5, 4, 3, 2, 1};
+  std::vector<int64_t> list1 = {1, 2, 3, 4, 5};
+  std::vector<int64_t> list2 = {5, 4, 3, 2, 1};
   AbstractBasePtr abs_t1 = FromValue(list1, true);
   AbstractBasePtr abs_t2 = FromValue(list2, true);
 
   AbstractBasePtr res_t1 = abs_t1->Join(abs_t2);
   ASSERT_EQ(res_t1, abs_t1);
 
-  abs_s1 = FromValue(1, false);
+  abs_s1 = FromValue(static_cast<int64_t>(1), false);
 
   AbstractBasePtr t1 = std::make_shared<AbstractTuple>(AbstractBasePtrList({abs_s1, abs_s_anything}));
   AbstractBasePtr t2 = std::make_shared<AbstractTuple>(AbstractBasePtrList({abs_s1, abs_s_anything}));
