+++ tests/ut/cpp/parallel/ops_info/onehot_info_test_axis_0.cc
@@ -54,7 +54,7 @@ void TestOneHotInfo2::SetUp() {
   g_device_manager = std::make_shared<DeviceManager>();
   g_device_manager->Init(dev_list, local_dev, stage_map, "hccl");
 
-  ValuePtr axis = MakeValue(std::int32_t(0));
+  ValuePtr axis = MakeValue(std::int64_t(0));
   std::unordered_map<std::string, ValuePtr> attr = {{"axis", axis}};
 
   Shapes inputs_shape = {{64}, {}, {}};
